Screen: Make textures functional on OpenGL-ES-1
Screen('Maketexture/DrawTexture/DrawTextures') as well as video capture
and movie playback works. Drawing into classic offscreen windows and
Screen('CopyWindow') work as well.

ES is limited to 8-Bit textures, or more specifically, to Luminance,
Luminance+Alpha, RGB or RGBA textures of unspecified precision. But
the assumption is that usually 8-Bit fixed point storage for 256 levels
per color component are supported by the underlying hardware.

This also enables use of non-power-of-two GL_TEXTURE_2D textures if
the GL_OES_texture_npot extension is available - Usually only on
ES-2, not ES-1 though.

+++ PsychSourceGL/Source/Common/Screen/PsychGLGlue.c
@@ -784,10 +784,10 @@ void PsychGLVertex4f(PsychWindowRecordType *windowRecord, float x, float y, floa
 void PsychGLColor4f(PsychWindowRecordType *windowRecord, float r, float g, float b, float a)
 {
     double colors[4];
-    colors[0] = (float) r;
-    colors[1] = (float) g;
-    colors[2] = (float) b;
-    colors[3] = (float) a;
+    colors[0] = r;
+    colors[1] = g;
+    colors[2] = b;
+    colors[3] = a;
     PsychSetArrayColor(windowRecord, 0, 4, colors, NULL);
 }
